feat: Allow banning users by replying to their message
Enhances the `/ban` command to provide a more intuitive user experience. Administrators can now ban a user by simply replying to one of their messages, removing the need to explicitly mention them.

When a user is banned using this reply method, the original message they sent is also deleted from the group. The command's help text is updated to reflect this new capability.